+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.15.0] - 2026-06-03
 
-###
+### Added
 - Added proper support for external listeners [#523](https://github.com/ie3-institute/simonaAPI/issues/523)
+- Added addon loader functionality [#528](https://github.com/ie3-institute/simonaAPI/issues/528)
 
 ### Changed
 - Adapted `EnergyBoundariesFlexOptions` [#489](https://github.com/ie3-institute/simonaAPI/issues/489)